feat(trading): update margin levels component
